client tests working

+++ stac_api/clients/postgres/transactions.py
@@ -5,14 +5,14 @@ import logging
 from typing import Dict, List, Optional, Type, Union
 
 import attr
+import sqlalchemy as sa
 from sqlalchemy import create_engine
-from sqlalchemy.orm import Session as SqlSession
 
+import psycopg2
 from stac_api import errors
 from stac_api.clients.base import BaseTransactionsClient, BulkTransactionsClient
 from stac_api.clients.postgres.session import Session
 from stac_api.models import database, schemas
-from stac_api.errors import DatabaseError, NotFoundError
 
 logger = logging.getLogger(__name__)
 
@@ -26,54 +26,81 @@ class TransactionsClient(BaseTransactionsClient):
     item_table: Type[database.Item] = attr.ib(default=database.Item)
 
     def create_item(self, model: schemas.Item, **kwargs) -> schemas.Item:
+        """create item"""
         data = self.item_table.from_schema(model)
-        with self.session.writer.context_session() as session:
-            session.add(data)
+        try:
+            with self.session.writer.context_session() as session:
+                session.add(data)
+        except sa.exc.IntegrityError as e:
+            if isinstance(e.orig, psycopg2.errors.UniqueViolation):
+                raise errors.ConflictError(f"Item {model.id} already exists")
         return model
 
     def create_collection(
         self, model: schemas.Collection, **kwargs
     ) -> schemas.Collection:
+        """create collection"""
         data = self.collection_table.from_schema(model)
-        with self.session.writer.context_session() as session:
-            session.add(data)
+        try:
+            with self.session.writer.context_session() as session:
+                session.add(data)
+        except sa.exc.IntegrityError as e:
+            if isinstance(e.orig, psycopg2.errors.UniqueViolation):
+                raise errors.ConflictError(f"Collection {model.id} already exists")
         return model
 
     def update_item(self, model: schemas.Item, **kwargs) -> schemas.Item:
+        """update item"""
         with self.session.reader.context_session() as session:
-            query = session.query(self.item_table).filter(self.item_table.id == model.id)
+            query = session.query(self.item_table).filter(
+                self.item_table.id == model.id
+            )
             if not query.scalar():
-                raise NotFoundError(f"Item {model.id} not found")
+                raise errors.NotFoundError(f"Item {model.id} not found")
             # SQLAlchemy orm updates don't seem to like geoalchemy types
             data = self.item_table.get_database_model(model)
             data.pop("geometry", None)
+            query.update(data)
         return model
 
     def update_collection(
         self, model: schemas.Collection, **kwargs
     ) -> schemas.Collection:
+        """update collection"""
         with self.session.reader.context_session() as session:
-            query = session.query(self.item_table).filter(self.item_table.id == model.id)
+            query = session.query(self.collection_table).filter(
+                self.collection_table.id == model.id
+            )
             if not query.scalar():
-                raise NotFoundError(f"Item {model.id} not found")
+                raise errors.NotFoundError(f"Item {model.id} not found")
             # SQLAlchemy orm updates don't seem to like geoalchemy types
             data = self.collection_table.get_database_model(model)
             data.pop("geometry", None)
+            query.update(data)
         return model
 
     def delete_item(self, id: str, **kwargs) -> schemas.Item:
+        """delete item"""
         with self.session.writer.context_session() as session:
             query = session.query(self.item_table).filter(self.item_table.id == id)
             data = query.first()
+            if not data:
+                raise errors.NotFoundError(f"Item {id} not found")
             query.delete()
         return data
 
     def delete_collection(self, id: str, **kwargs) -> schemas.Item:
+        """delete collection"""
         with self.session.writer.context_session() as session:
-            query = session.query(self.collection_table).filter(self.item_table.id == id)
+            query = session.query(self.collection_table).filter(
+                self.collection_table.id == id
+            )
             data = query.first()
+            if not data:
+                raise errors.NotFoundError(f"Collection {id} not found")
             query.delete()
         return data
+
 
 @attr.s
 class PostgresBulkTransactions(BulkTransactionsClient):
